changed storage type in cash box
